+++ interfaces/microphysics.F90
@@ -1,10 +1,14 @@
 module microphysics_module
 
   use network
+#ifdef MICROPHYSICS_FORT_EOS
   use eos_module, only : eos_init
+#endif
 #ifdef REACTIONS
+#ifndef TRUE_SDC
   use actual_rhs_module, only : actual_rhs_init
-#ifndef SIMPLIFIED_SDC
+#endif
+#ifdef STRANG
   use actual_burner_module, only : actual_burner_init
 #endif
 #endif
@@ -34,6 +38,7 @@ contains
 
     call network_init()
 
+#ifdef MICROPHYSICS_FORT_EOS
     if (present(small_temp) .and. present(small_dens)) then
        call eos_init(small_temp=small_temp, small_dens=small_dens)
     else if (present(small_temp)) then
@@ -43,10 +48,13 @@ contains
     else
        call eos_init()
     endif
+#endif
 
 #ifdef REACTIONS
+#ifndef TRUE_SDC
     call actual_rhs_init()
-#ifndef SIMPLIFIED_SDC
+#endif
+#ifdef STRANG
     call actual_burner_init()
 #endif
 #endif
@@ -59,12 +67,17 @@ contains
 
   subroutine microphysics_finalize() bind(C, name="microphysics_finalize")
 
+#ifdef MICROPHYSICS_FORT_EOS
     use eos_module, only: eos_finalize
+#endif
 #ifdef USE_SCREENING
     use screening_module, only: screening_finalize
     call screening_finalize()
 #endif
+
+#ifdef MICROPHYSICS_FORT_EOS
     call eos_finalize()
+#endif
     call network_finalize()
 
   end subroutine microphysics_finalize