+++ networks/aprox19/actual_burner.F90
@@ -17,7 +17,7 @@ contains
   subroutine actual_burner(state_in, state_out, dt, time)
 
     use integrator_module, only: integrator
-
+    use extern_probin_module, only : eta
     implicit none
 
     type (burn_t),       intent(in   ) :: state_in
@@ -56,6 +56,13 @@ contains
        call nse_interp(state_in % T, state_in % rho, state_in % aux(iye), &
                        abar_out, dq_out, dyedt, state_out % xn(:))
 
+       ! update Ye
+       state_out % aux(iye) = state_in % aux(iye) + dt * dyedt
+
+       ! now get the composition from the table using the updated Ye
+       call nse_interp(state_in % T, state_in % rho, state_out % aux(iye), &
+                       abar_out, dq_out, dyedt, state_out % xn(:))
+
        state_out % success = .true.
        state_out % n_rhs = 0
        state_out % n_jac = 0
@@ -66,7 +73,7 @@ contains
        deltaq = dq_out - state_in % aux(ibea)
 
        ! under-relaxation / inertia (see Ma et el. 2013)
-       deltaq = 0.3_rt * deltaq
+       deltaq = eta * deltaq
 
        state_out % aux(ibea) = state_in % aux(ibea) + deltaq
 
@@ -75,7 +82,6 @@ contains
 
        state_out % e = enuc + state_in % e
 
-       state_out % aux(iye) = state_in % aux(iye) + dt * dyedt
        state_out % aux(iabar) = abar_out
 
     end if