+++ sphinx_docs/source/conf.py
@@ -62,6 +62,9 @@ breathe_projects = {
 
 breathe_default_project = "microphysics"
 
+# bibtex
+bibtex_bibfiles = ["refs.bib"]
+
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']