+++ unit_test/test_eos/variables.cpp
@@ -55,62 +55,68 @@ plot_t init_variables() {
   p.idpde = p.next_index(1);
   p.idpdre = p.next_index(1);
 
-  p.names.resize(p.n_plot_comps);
-
-  p.names[p.irho] = "density";
-  p.names[p.itemp] = "temperature";
-  p.names[p.ih] = "specific_enthalpy";
-  p.names[p.ie] = "specific_energy";
-  p.names[p.ip] = "pressure";
-  p.names[p.is] = "specific_entropy";
-  for (int n = 0; n < NumSpec; n++) {
-    p.names[p.ispec + n] = "X_" + spec_names_cxx[n];
-  }
-
-  p.names[p.ierr_T_eos_rh] = "err_T_eos_rh";
-  p.names[p.ierr_rho_eos_tp] = "err_rho_eos_tp";
-  p.names[p.ierr_T_eos_rp] = "err_T_eos_rp";
-  p.names[p.ierr_T_eos_re] = "err_T_eos_re";
-  p.names[p.ierr_rho_eos_ps] = "err_rho_eos_ps";
-  p.names[p.ierr_T_eos_ps] = "err_T_eos_ps";
-  p.names[p.ierr_rho_eos_ph] = "err_rho_eos_ph";
-  p.names[p.ierr_T_eos_ph] = "err_T_eos_ph";
-  p.names[p.ierr_rho_eos_th] = "err_rho_eos_th";
-
-  p.names[p.icv] = "c_v";
-  p.names[p.icp] = "c_p";
-  p.names[p.ine] = "n_e";
-  p.names[p.inp] = "n_p";
-  p.names[p.ieta] = "eta";
-  p.names[p.ipele] = "p_ele";
-  p.names[p.ippos] = "p_pos";
-  p.names[p.imu] = "mu";
-  p.names[p.imue] = "mu_e";
-  p.names[p.iye] = "Y_e";
-  p.names[p.idpdt] = "dp_dT";
-  p.names[p.idpdr] = "dp_drho";
-  p.names[p.idedt] = "de_dT";
-  p.names[p.idedr] = "de_drho";
-  p.names[p.idsdt] = "ds_dT";
-  p.names[p.idsdr] = "ds_drho";
-  p.names[p.idhdt] = "dh_dT";
-  p.names[p.idhdr] = "dh_drho";
-  for (int n = 0; n < NumSpec; n++) {
-    p.names[p.idpdx + n] = "dp_dX_" + spec_names_cxx[n];
-    p.names[p.idedx + n] = "de_dX_" + spec_names_cxx[n];
-    p.names[p.idhdx + n] = "dh_dX_" + spec_names_cxx[n];
-  }
-  p.names[p.igam1] = "Gamma_1";
-  p.names[p.ics] = "soundspeed";
-  p.names[p.iabar] = "Abar";
-  p.names[p.izbar] = "Zbar";
-  p.names[p.idpda] = "dp_dA";
-  p.names[p.idpdz] = "dp_dZ";
-  p.names[p.ideda] = "de_dA";
-  p.names[p.idedz] = "de_dZ";
-  p.names[p.idpde] = "dp_de_rho";
-  p.names[p.idpdre] = "dp_drho_e";
-
   return p;
 }
+
+
+void get_varnames(const plot_t p, amrex::Vector<std::string>& names) {
+
+  names.resize(p.n_plot_comps);
+
+  names[p.irho] = "density";
+  names[p.itemp] = "temperature";
+  names[p.ih] = "specific_enthalpy";
+  names[p.ie] = "specific_energy";
+  names[p.ip] = "pressure";
+  names[p.is] = "specific_entropy";
+  for (int n = 0; n < NumSpec; n++) {
+    names[p.ispec + n] = "X_" + spec_names_cxx[n];
+  }
+
+  names[p.ierr_T_eos_rh] = "err_T_eos_rh";
+  names[p.ierr_rho_eos_tp] = "err_rho_eos_tp";
+  names[p.ierr_T_eos_rp] = "err_T_eos_rp";
+  names[p.ierr_T_eos_re] = "err_T_eos_re";
+  names[p.ierr_rho_eos_ps] = "err_rho_eos_ps";
+  names[p.ierr_T_eos_ps] = "err_T_eos_ps";
+  names[p.ierr_rho_eos_ph] = "err_rho_eos_ph";
+  names[p.ierr_T_eos_ph] = "err_T_eos_ph";
+  names[p.ierr_rho_eos_th] = "err_rho_eos_th";
+
+  names[p.icv] = "c_v";
+  names[p.icp] = "c_p";
+  names[p.ine] = "n_e";
+  names[p.inp] = "n_p";
+  names[p.ieta] = "eta";
+  names[p.ipele] = "p_ele";
+  names[p.ippos] = "p_pos";
+  names[p.imu] = "mu";
+  names[p.imue] = "mu_e";
+  names[p.iye] = "Y_e";
+  names[p.idpdt] = "dp_dT";
+  names[p.idpdr] = "dp_drho";
+  names[p.idedt] = "de_dT";
+  names[p.idedr] = "de_drho";
+  names[p.idsdt] = "ds_dT";
+  names[p.idsdr] = "ds_drho";
+  names[p.idhdt] = "dh_dT";
+  names[p.idhdr] = "dh_drho";
+  for (int n = 0; n < NumSpec; n++) {
+    names[p.idpdx + n] = "dp_dX_" + spec_names_cxx[n];
+    names[p.idedx + n] = "de_dX_" + spec_names_cxx[n];
+    names[p.idhdx + n] = "dh_dX_" + spec_names_cxx[n];
+  }
+  names[p.igam1] = "Gamma_1";
+  names[p.ics] = "soundspeed";
+  names[p.iabar] = "Abar";
+  names[p.izbar] = "Zbar";
+  names[p.idpda] = "dp_dA";
+  names[p.idpdz] = "dp_dZ";
+  names[p.ideda] = "de_dA";
+  names[p.idedz] = "de_dZ";
+  names[p.idpde] = "dp_de_rho";
+  names[p.idpdre] = "dp_drho_e";
+
+}
+