+++ unit_test/test_parameters/test_parameters.F90
@@ -7,6 +7,10 @@ subroutine do_f90_parameters() bind(C, name="do_f90_parameters")
   !$gpu
 
   print *, "in Fortran"
-  print *, "eos_input_is_constant = ", eos_input_is_constant
+  print *, "  eos_input_is_constant = ", eos_input_is_constant
+  print *, "  test_string = ", test_string
+  print *, "  dens_min = ", dens_min
+  print *, "  nonaka_file = ", nonaka_file
+  print *, " "
 
 end subroutine do_f90_parameters